+++ app/hooks/use-tender-context.ts
@@ -1,11 +1,28 @@
 import { useState, useEffect } from "react";
-import type { Market, TenderContext } from "~/types/global";
+import { Market, TenderContext, TokenName } from "~/types/global";
 import { hooks as Web3Hooks } from "~/connectors/meta-mask";
 import { useTokenPairs } from "./use-token-pairs";
 import { useOnSupportedNetwork } from "./use-on-supported-network";
 import { useNetworkData } from "./use-network-data";
-import { SUPPORTED_TOKENS } from "../components/swap-table";
 import { useMarkets } from "./use-markets";
+
+// TODO: Move to config
+export const SUPPORTED_TOKENS = [
+  TokenName.AAVE,
+  TokenName.BAT,
+  TokenName.DAI,
+  TokenName.ETH,
+  TokenName.LINK,
+  TokenName.PAX,
+  TokenName.SUSHI,
+  TokenName.TUSD,
+  TokenName.UNI,
+  TokenName.USDC,
+  TokenName.USDT,
+  TokenName.WBTC,
+  TokenName.YFI,
+  TokenName.ZRX,
+];
 
 export function useTenderContext() {
   let [tenderContext, setTenderContext] = useState<TenderContext | null>();
